du for subdirs
There is a trouble with du when you whant to count all subdirs
and subfiles.

Due to *sh expansion logic ./* is all non hidden files.

So when you whant to calculate everythins you could do the following
du -sm .* *
But that expand to . .. and all subdirs. And du with -s stop after count
of .

So right way to deal with that is du -m -a --with-depth 1 .